+++ types/cloudflare.d.ts
@@ -1,3 +1,0 @@
-declare interface Fetcher {
-  fetch(input: RequestInfo | URL, init?: RequestInit): Promise<Response>;
-}